+++ includes/features/class-wpcli.php
@@ -29,6 +29,149 @@ use APCuManager\System\Markdown;
 class Wpcli {
 
 	/**
+	 * List of exit codes.
+	 *
+	 * @since    2.0.0
+	 * @var array $exit_codes Exit codes.
+	 */
+	private static $exit_codes = [
+		0   => 'operation successful.',
+		1   => 'unrecognized setting.',
+		2   => 'unrecognized action.',
+		3   => 'analytics are disabled.',
+		255 => 'unknown error.',
+	];
+
+	/**
+	 * Write ids as clean stdout.
+	 *
+	 * @param   array   $ids   The ids.
+	 * @param   string  $field  Optional. The field to output.
+	 * @since   2.0.0
+	 */
+	private static function write_ids( $ids, $field = '' ) {
+		$result = '';
+		$last   = end( $ids );
+		foreach ( $ids as $key => $id ) {
+			if ( '' === $field ) {
+				$result .= $key;
+			} else {
+				$result .= $id[$field];
+			}
+			if ( $id !== $last ) {
+				$result .= ' ';
+			}
+		}
+		// phpcs:ignore
+		fwrite( STDOUT, $result );
+	}
+
+	/**
+	 * Write an error.
+	 *
+	 * @param   integer  $code      Optional. The error code.
+	 * @param   boolean  $stdout    Optional. Clean stdout output.
+	 * @since   2.0.0
+	 */
+	private static function error( $code = 255, $stdout = false ) {
+		if ( \WP_CLI\Utils\isPiped() ) {
+			// phpcs:ignore
+			fwrite( STDOUT, '' );
+			// phpcs:ignore
+			exit( $code );
+		} elseif ( $stdout ) {
+			// phpcs:ignore
+			fwrite( STDERR, ucfirst( self::$exit_codes[ $code ] ) );
+			// phpcs:ignore
+			exit( $code );
+		} else {
+			\WP_CLI::error( self::$exit_codes[ $code ] );
+		}
+	}
+
+	/**
+	 * Write a warning.
+	 *
+	 * @param   string   $msg       The message.
+	 * @param   string   $result    Optional. The result.
+	 * @param   boolean  $stdout    Optional. Clean stdout output.
+	 * @since   2.0.0
+	 */
+	private static function warning( $msg, $result = '', $stdout = false ) {
+		if ( \WP_CLI\Utils\isPiped() || $stdout ) {
+			// phpcs:ignore
+			fwrite( STDOUT, $result );
+		} else {
+			\WP_CLI::warning( $msg );
+		}
+	}
+
+	/**
+	 * Write a success.
+	 *
+	 * @param   string   $msg       The message.
+	 * @param   string   $result    Optional. The result.
+	 * @param   boolean  $stdout    Optional. Clean stdout output.
+	 * @since   2.0.0
+	 */
+	private static function success( $msg, $result = '', $stdout = false ) {
+		if ( \WP_CLI\Utils\isPiped() || $stdout ) {
+			// phpcs:ignore
+			fwrite( STDOUT, $result );
+		} else {
+			\WP_CLI::success( $msg );
+		}
+	}
+
+	/**
+	 * Write a wimple line.
+	 *
+	 * @param   string   $msg       The message.
+	 * @param   string   $result    Optional. The result.
+	 * @param   boolean  $stdout    Optional. Clean stdout output.
+	 * @since   2.0.0
+	 */
+	private static function line( $msg, $result = '', $stdout = false ) {
+		if ( \WP_CLI\Utils\isPiped() || $stdout ) {
+			// phpcs:ignore
+			fwrite( STDOUT, $result );
+		} else {
+			\WP_CLI::line( $msg );
+		}
+	}
+
+	/**
+	 * Write a wimple log line.
+	 *
+	 * @param   string   $msg       The message.
+	 * @param   boolean  $stdout    Optional. Clean stdout output.
+	 * @since   2.0.0
+	 */
+	private static function log( $msg, $stdout = false ) {
+		if ( ! \WP_CLI\Utils\isPiped() && ! $stdout ) {
+			\WP_CLI::log( $msg );
+		}
+	}
+
+	/**
+	 * Get params from command line.
+	 *
+	 * @param   array   $args   The command line parameters.
+	 * @return  array The true parameters.
+	 * @since   2.0.0
+	 */
+	private static function get_params( $args ) {
+		$result = '';
+		if ( array_key_exists( 'settings', $args ) ) {
+			$result = \json_decode( $args['settings'], true );
+		}
+		if ( ! $result || ! is_array( $result ) ) {
+			$result = [];
+		}
+		return $result;
+	}
+
+	/**
 	 * Get APCu Manager details and operation modes.
 	 *
 	 * ## EXAMPLES
@@ -76,6 +219,9 @@ class Wpcli {
 	 * [--yes]
 	 * : Answer yes to the confirmation message, if any.
 	 *
+	 * [--stdout]
+	 * : Use clean STDOUT output to use results in scripts. Unnecessary when piping commands because piping is detected by APCu Manager.
+	 *
 	 * ## EXAMPLES
 	 *
 	 * wp apcu settings enable analytics
@@ -86,6 +232,7 @@ class Wpcli {
 	 *
 	 */
 	public static function settings( $args, $assoc_args ) {
+		$stdout  = \WP_CLI\Utils\get_flag_value( $assoc_args, 'stdout', false );
 		$action  = isset( $args[0] ) ? (string) $args[0] : '';
 		$setting = isset( $args[1] ) ? (string) $args[1] : '';
 		switch ( $action ) {
@@ -93,14 +240,14 @@ class Wpcli {
 				switch ( $setting ) {
 					case 'analytics':
 						Option::network_set( 'analytics', true );
-						\WP_CLI::success( 'analytics are now activated.' );
+						self::success( 'analytics are now activated.', '', $stdout );
 						break;
 					case 'gc':
 						Option::network_set( 'gc', true );
-						\WP_CLI::success( 'garbage collection is now activated.' );
+						self::success( 'garbage collection is now activated.', '', $stdout );
 						break;
 					default:
-						\WP_CLI::error( 'unrecognized setting.' );
+						self::error( 1, $stdout );
 				}
 				break;
 			case 'disable':
@@ -108,19 +255,19 @@ class Wpcli {
 					case 'analytics':
 						\WP_CLI::confirm( 'Are you sure you want to deactivate analytics?', $assoc_args );
 						Option::network_set( 'analytics', false );
-						\WP_CLI::success( 'analytics are now deactivated.' );
+						self::success( 'analytics are now deactivated.', '', $stdout );
 						break;
 					case 'gc':
 						\WP_CLI::confirm( 'Are you sure you want to deactivate garbage collection?', $assoc_args );
 						Option::network_set( 'gc', false );
-						\WP_CLI::success( 'garbage collection is now deactivated.' );
+						self::success( 'garbage collection is now deactivated.', '', $stdout );
 						break;
 					default:
-						\WP_CLI::error( 'unrecognized setting.' );
+						self::error( 1, $stdout );
 				}
 				break;
 			default:
-				\WP_CLI::error( 'unrecognized action.' );
+				self::error( 2, $stdout );
 		}
 	}
 
@@ -139,6 +286,9 @@ class Wpcli {
 	 *  - count
 	 * ---
 	 *
+	 * [--stdout]
+	 * : Use clean STDOUT output to use results in scripts. Unnecessary when piping commands because piping is detected by APCu Manager.
+	 *
 	 * ## EXAMPLES
 	 *
 	 * wp apcu analytics
@@ -148,8 +298,9 @@ class Wpcli {
 	 *
 	 */
 	public static function analytics( $args, $assoc_args ) {
+		$stdout = \WP_CLI\Utils\get_flag_value( $assoc_args, 'stdout', false );
 		if ( ! Option::network_get( 'analytics' ) ) {
-			\WP_CLI::error( 'analytics are disabled.' );
+			self::error( 3, $stdout );
 		}
 		$analytics = Analytics::get_status_kpi_collection();
 		$result    = [];
@@ -168,11 +319,65 @@ class Wpcli {
 				$result[]          = $item;
 			}
 		}
-		$format = isset( $assoc_args['format'] ) ? $assoc_args['format'] : 'table';
+		$format = \WP_CLI\Utils\get_flag_value( $assoc_args, 'format', 'table' );
 		if ( 'json' === $format ) {
-			\WP_CLI::line( wp_json_encode( $analytics ) );
+			self::line( wp_json_encode( $analytics ), wp_json_encode( $analytics ), $stdout );
 		} else {
 			\WP_CLI\Utils\format_items( $assoc_args['format'], $result, [ 'kpi', 'description', 'value', 'ratio', 'variation' ] );
+		}
+	}
+
+	/**
+	 * Get information on exit codes.
+	 *
+	 * ## OPTIONS
+	 *
+	 * <list>
+	 * : The action to take.
+	 * ---
+	 * options:
+	 *  - list
+	 * ---
+	 *
+	 * [--format=<format>]
+	 * : Allows overriding the output of the command when listing exit codes.
+	 * ---
+	 * default: table
+	 * options:
+	 *  - table
+	 *  - json
+	 *  - csv
+	 *  - yaml
+	 *  - ids
+	 *  - count
+	 * ---
+	 *
+	 * ## EXAMPLES
+	 *
+	 * Lists available exit codes:
+	 * + wp apcu exitcode list
+	 * + wp apcu exitcode list --format=json
+	 *
+	 *
+	 *   === For other examples and recipes, visit https://github.com/Pierre-Lannoy/wp-traffic/blob/master/WP-CLI.md ===
+	 *
+	 */
+	public static function exitcode( $args, $assoc_args ) {
+		$stdout = \WP_CLI\Utils\get_flag_value( $assoc_args, 'stdout', false );
+		$format = \WP_CLI\Utils\get_flag_value( $assoc_args, 'format', 'table' );
+		$action = isset( $args[0] ) ? $args[0] : 'list';
+		$codes  = [];
+		foreach ( self::$exit_codes as $key => $msg ) {
+			$codes[ $key ] = [ 'code' => $key, 'meaning' => ucfirst( $msg ) ];
+		}
+		switch ( $action ) {
+			case 'list':
+				if ( 'ids' === $format ) {
+					self::write_ids( $codes );
+				} else {
+					\WP_CLI\Utils\format_items( $format, $codes, [ 'code', 'meaning' ] );
+				}
+				break;
 		}
 	}
 
@@ -197,4 +402,5 @@ if ( defined( 'WP_CLI' ) && WP_CLI ) {
 	\WP_CLI::add_command( 'apcu status', [ Wpcli::class, 'status' ] );
 	\WP_CLI::add_command( 'apcu analytics', [ Wpcli::class, 'analytics' ] );
 	\WP_CLI::add_command( 'apcu settings', [ Wpcli::class, 'settings' ] );
+	\WP_CLI::add_command( 'apcu exitcode', [ Wpcli::class, 'exitcode' ] );
 }