+++ CHANGELOG.md
@@ -3,10 +3,14 @@ All notable changes to **APCu Manager** are documented in this *changelog*.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and **APCu Manager** adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased - will be 2.0.1]
+## [Unreleased - will be 2.1.0]
+
+### Added
+- [WP-CLI] New command to discover exit codes: see `wp help apcu exitcode` for details.
 
 ### Changed
 - Improved IP detection  (thanks to [Ludovic Riaudel](https://github.com/lriaudel)).
+- [WP-CLI] `wp apcu settings` and `wp apcu analytics` now support `--stdout` flag.
 - [WP-CLI] Improved documentation.
 
 ### Fixed